+++ src/portioncollection.cpp
@@ -3,24 +3,24 @@
 
 PortionCollection::PortionCollection() {}
 
-void PortionCollection::addPortion( QString foodName, uint xValue,
-									uint yValue ) {
+void PortionCollection::addPortion( QString foodName, double xValue,
+									double yValue ) {
 
 	m_portions.append( Portion( foodName, xValue, yValue ) );
 }
 
 const QVector<Portion> &PortionCollection::getPortions() { return m_portions; }
 
-uint PortionCollection::getMaxXValue() {
-	uint maxValue = 0;
+double PortionCollection::getMaxXValue() {
+	double maxValue = 0;
 	for ( Portion p : m_portions ) {
 		maxValue = qMax( maxValue, p.m_xValue );
 	}
 	return maxValue;
 }
 
-uint PortionCollection::getTotalYValue() {
-	uint totalValue = 0;
+double PortionCollection::getTotalYValue() {
+	double totalValue = 0;
 	for ( Portion p : m_portions ) {
 		totalValue += p.m_yValue;
 	}
